Reorganize styles folder structure
