+++ client/src/components/main/NotesPanel.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ClipLoader } from "react-spinners";
 import { useQuery } from "@tanstack/react-query";
 
-import styles from "../../styles/NotesPanel.module.css";
+import styles from "../../styles/main/NotesPanel.module.css";
 import { CreateNote } from "../secondary/CreateNote";
 import { Notes } from "../secondary/Notes";
 import { readNotesRequest } from "../../API/readNotesRequest";
