+++ client/src/components/utility/ConfirmationPopup.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import styles from "../../styles/utility/ConfirmationPopup.module.css";
+
+export const ConfirmationPopup = () => {
+  return <div>ConfirmationPopup</div>;
+};
