+++ client/src/styles/main/NotesPanel.module.css
